Moved graphql retrivals to database file
